Image viewer Update
Added small quality of life change where user can now view the table in Foster 2006 which gives the terms used throughout the paper, and the ability to close and open at will.
NOTE: The PNG file must remain in the directory for it to be loaded.

This can be used as a template to open any PNG in a window for future reference.